automatically backed up by learn

+++ index.js
@@ -26,5 +26,7 @@ function makeInt(n){
 }
 
 function preserveDecimal(n){
-  return parseInt(parseFloat(n),10)
+  return parseFloat(parseInt(n,10))
 }
+
+
